thcrap_tasofro: pl: fix child versions of parseCommand not being called
Commit 068829b changed a parameter for AText::parseCommand from an int to a
size_t. But it didn't apply the same changes to StoryText::parseCommand and
Th155StoryText::parseCommand, which were overriding that function.
Because of that, the 2 functions from the child classes were no longer
overriding the one from the parent class.

Fixes an issue in th145 where only the last line of dialog balloons was
displayed, and may fix #286

+++ thcrap_tasofro/src/pl.cpp
@@ -493,7 +493,7 @@ bool TasofroPl::AText::parseCommand(json_t *patch, size_t json_line_num)
 	return false;
 }
 
-bool TasofroPl::StoryText::parseCommand(json_t *patch, int json_line_num)
+bool TasofroPl::StoryText::parseCommand(json_t *patch, size_t json_line_num)
 {
 	bool ret = this->AText::parseCommand(patch, json_line_num);
 	if (ret == false) {
@@ -504,7 +504,7 @@ bool TasofroPl::StoryText::parseCommand(json_t *patch, int json_line_num)
 	return ret;
 }
 
-bool TasofroPl::Th155StoryText::parseCommand(json_t *patch, int json_line_num)
+bool TasofroPl::Th155StoryText::parseCommand(json_t *patch, size_t json_line_num)
 {
 	// Overwrite the StoryText overwrite and go back to the standard behavior
 	return this->AText::parseCommand(patch, json_line_num);
